requirements.txt
aniso8601==2.0.1
astroid==1.6.1
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.5.1
coveralls==1.2.0
docopt==0.6.2
Flask==0.12.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.6
flask-restful-swagger==0.19
flask-swagger==0.2.13
gunicorn==19.7.1
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
nosexcover==1.0.11
pluggy==0.6.0
py==1.5.2
PyJWT==1.4.2
pylint==1.8.2
pytz==2018.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
